Code optimizations, debugging
Added event functionality to sprites (the end of an animation can trigger a callback).  Entities can now have their own version of the same sprite (tweaked the Sprite object to allow for cloneable pooling). Added a resource handler for loading sprites - and audio in the future. The way lanes are handled is now more flexible (has a separate object).  Changed tutorial code to be more flexible in addition to solving various issues with the timing of tutorial popups.  Adjusted other code snippets in accordance with the above changes.